dependency_builder/pyproject.toml
[tool.poetry]
name = "dependency_builder"
version = "0.1.0"
description = ""
authors = ["Mike Hukiewitz <mike.hukiewitz@robotter.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
aleph-sdk-python = "^0.6.0"
fastapi = "^0.96.0"
python-multipart = "^0.0.6"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
uvicorn = {extras = ["standard"], version = "^0.22.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
